requirements.txt
aiohttp~=3.7.4.post0
matplotlib
asyncpg~=0.24.0
aiogram~=2.14.3
python-dotenv~=0.19.0
aioredis~=1.3.1
dataclasses_json~=0.5.5
python-dateutil~=2.8.2
glqiwiapi~=1.0.3
APScheduler~=3.7.0
marshmallow~=3.13.0